Work5.0

+++ MetricsAgent/DAL/Helpers/DateTimeHandler.cs
@@ -10,7 +10,7 @@ namespace MetricsAgent.DAL
         public override void SetValue(IDbDataParameter parameter, DateTimeOffset value)=>parameter.Value = value.ToUnixTimeSeconds();
   
 
-        public override DateTimeOffset Parse(object value)=>  DateTimeOffset.FromUnixTimeSeconds((long) value);
+        public override DateTimeOffset Parse(object value)=>  DateTimeOffset.FromUnixTimeSeconds((int) value);
       
     }
 }
